app - transaction simple test added, id type long > string

+++ src/main/kotlin/com/kyu9/accountbook/application/TransactionService.kt
@@ -38,7 +38,7 @@ class TransactionService(
     }
 
     fun getSingleTransaction(utid: String): GetSingleTransResponseDto {
-        return transactionRepoImpl.getEntityWithId(utid.toLong()).let{
+        return transactionRepoImpl.getEntityWithId(utid).let{
             GetSingleTransResponseDto(
                 utid = it.id,
                 amount = it.amount.toInt(),
